Gunnar: -finite width tools         -fixing generation of composite fermion state
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@970 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnLattice.cc
@@ -609,7 +609,7 @@ void BosonOnLattice::ListQuantumNumbers(int index, int *quantumNumbers, double &
   for (int q=0; q<=TemporaryStateHighestBit; ++q)
     if (this->TemporaryState[q]>0)
       {		
-	for (unsigned int l=this->TemporaryState[q];l>0; --l)
+	for (unsigned long l=this->TemporaryState[q];l>0; --l)
 	  {
 	    normalization*=l;
 	    quantumNumbers[NbrQ]=q;
